pyproject.toml
[project]
name = "python-assignments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ALI-web242", email = "alinaqi1351@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.4",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "streamlit>=1.43.2",
]

[project.scripts]
python-assignments = "python_assignments:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
